airbyte-integrations/connectors/source-generic-http/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-generic-http"
description = "Source implementation for generic-http."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/generic-http"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_generic_http" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
airbyte-cdk = "^0"
pydash = "^8.0.0"

[tool.poetry.scripts]
source-generic-http = "source_generic_http.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"

